Split Ouroboros.Consensus.NodeId out of Ouroboros.Consensus.Node
The rest of Ouroboros.Consensus.Node does not need the NodeId and
almost all modules that import Ouroboros.Consensus.Node only actually
use the things from Ouroboros.Consensus.NodeId.

So since they're essentially completely independent then have them live
in separate modules and that makes it much clearer which things actually
depend on Ouroboros.Consensus.Node.
