+++ ouroboros-consensus/demo-playground/Topology.hs
@@ -12,7 +12,7 @@ import           Data.Map.Strict (Map)
 import qualified Data.Map.Strict as M
 import           Data.String.Conv (toS)
 
-import           Ouroboros.Consensus.Node (NodeId(..))
+import           Ouroboros.Consensus.NodeId (NodeId(..))
 
 -- | A data structure bundling together a node identifier and the path to
 -- the topology file.
