+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -82,6 +82,7 @@ library
                        Ouroboros.Consensus.Mempool.API
                        Ouroboros.Consensus.Mempool.Impl
                        Ouroboros.Consensus.Node
+                       Ouroboros.Consensus.NodeId
                        Ouroboros.Consensus.Protocol.Abstract
                        Ouroboros.Consensus.Protocol.BFT
                        Ouroboros.Consensus.Protocol.ExtNodeConfig
