+++ ouroboros-consensus/src/Ouroboros/Consensus/Demo/Ledger/Mock/Praos.hs
@@ -10,7 +10,7 @@ import           Ouroboros.Consensus.Crypto.VRF
 import           Ouroboros.Consensus.Demo.Run
 import           Ouroboros.Consensus.Ledger.Extended
 import           Ouroboros.Consensus.Ledger.Mock
-import           Ouroboros.Consensus.Node (CoreNodeId (..), NodeId (..))
+import           Ouroboros.Consensus.NodeId (CoreNodeId (..), NodeId (..))
 import           Ouroboros.Consensus.Protocol.ExtNodeConfig
 import           Ouroboros.Consensus.Protocol.Praos
