+++ ouroboros-consensus/test-consensus/Test/Consensus/ChainSyncClient.hs
@@ -49,8 +49,7 @@ import           Ouroboros.Consensus.BlockchainTime
 import           Ouroboros.Consensus.ChainSyncClient
 import           Ouroboros.Consensus.Crypto.DSIGN.Mock
 import           Ouroboros.Consensus.Ledger.Extended hiding (ledgerState)
-import           Ouroboros.Consensus.Node (CoreNodeId (..), NodeId (..),
-                     fromCoreNodeId)
+import           Ouroboros.Consensus.NodeId
 import           Ouroboros.Consensus.Protocol.Abstract
 import           Ouroboros.Consensus.Protocol.BFT
 import           Ouroboros.Consensus.Util (whenJust)
